toản comit khung project

+++ duantn_th/src/main/java/fpt/edu/duantn_th/entity/DiaChi.java
@@ -38,7 +38,7 @@ public class DiaChi {
     private String quocgia;
 
     @Column(name = "trangthai")
-    private Long trangthai;
+    private Integer trangthai;
 
     @ManyToOne
     @JoinColumn(name = "idtk")
