+++ duantn_th/src/main/java/fpt/edu/duantn_th/entity/DonHang.java
@@ -77,7 +77,7 @@ public class DonHang {
     private Double thanhtien;
 
     @Column(name = "trangthai")
-    private Long trangthai;
+    private Integer trangthai;
 
     @Column(name = "qrcode")
     private String qrcode;
